[Refact] ♻ Move commands & queries events
command 와 query 이벤트 자체는 도메인 주체인 것 같아서 옮겼습니다.
