+++ src/main/java/bboxx/application/controller/MemberController.java
@@ -6,10 +6,10 @@ import bboxx.application.controller.dto.response.NicknameGenerationApiData;
 import bboxx.application.security.AuthUserDetail;
 import bboxx.application.service.member.MemberCommandService;
 import bboxx.application.service.member.MemberQueryService;
-import bboxx.application.service.member.command.UpdateMemberCommand;
-import bboxx.application.service.member.query.GetMemberOneQuery;
+import bboxx.domain.member.command.UpdateMemberCommand;
+import bboxx.domain.member.query.GetMemberOneQuery;
 import bboxx.domain.member.commandmodel.NicknameGenerator;
-import bboxx.domain.member.readmodel.MemberView;
+import bboxx.domain.member.querymodel.MemberView;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
