+++ src/main/java/bboxx/domain/member/querymodel/MemberView.java
@@ -1,4 +1,4 @@
-package bboxx.domain.member.readmodel;
+package bboxx.domain.member.querymodel;
 
 import bboxx.domain.member.Member;
 import bboxx.domain.member.MemberState;
