+++ src/test/java/bboxx/BboxxBackendApplicationTests.java
@@ -3,11 +3,11 @@ package bboxx;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
-class BboxxBackendApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
+//@SpringBootTest
+//class BboxxBackendApplicationTests {
+//
+//	@Test
+//	void contextLoads() {
+//	}
+//
+//}
